+++ src/heat_helper/yeargroup.py
@@ -14,8 +14,8 @@ def clean_year_group(year_group: str | int, errors: str = "raise") -> str | None
         errors: default = 'raise' which raises all errors. 'ignore' returns original value, 'coerce' returns None.
 
     Raises:
-        InvalidYearGroupError: Raised when the year group input cannot be parsed or is out of range.
-        SchoolYearError: Raised when year_group is not a valid string.
+        InvalidYearGroupError: Raised when `year group` input cannot be parsed or is out of range.
+        TypeError: Raised when `year group` input cannot be parsed to a valid int.
 
     Returns:
         Cleaned year group in the format 'Year i'.
@@ -23,7 +23,7 @@ def clean_year_group(year_group: str | int, errors: str = "raise") -> str | None
     try:
         y_num = _parse_year_group_to_int(year_group)
         return "Reception" if y_num == 0 else f"Year {y_num}"
-    except (InvalidYearGroupError, TypeError, ValueError):
+    except (InvalidYearGroupError, TypeError):
         if errors == "ignore":
             return str(year_group)
         if errors == "coerce":
@@ -41,12 +41,14 @@ def calculate_year_group_from_date(
     Args:
         input_date: Date of birth you wish to know the school year for.
         start_of_academic_year (optional): The school year in which you want to calculate the year group for. Allows you to calculate a year group for any academic year not just current e.g. for 2025/26 school year enter 2025. Default is start of current academic year.
+        errors (optional): default = 'raise' which raises all errors. 'ignore' returns None.
 
     Raises:
         TypeError: Raised if input_date is not a date.
+        InvalidYearGroupError: Raised when the calculated year group is out of range (greater than Year 13).
 
     Returns:
-        Returns 'Year i' or 'Reception' or 'Student too young for school'.
+        Returns 'Year i', 'Reception', 'Student too young for school' or None if errors='ignore'.
     """
     try:
         if not isinstance(input_date, date):
@@ -64,7 +66,7 @@ def calculate_year_group_from_date(
             raise InvalidYearGroupError(f"Year {year_group}")
         else:
             return f"Year {year_group}"
-    except:
+    except (TypeError, InvalidYearGroupError):
         if errors == "ignore":
             return None
         raise